+++ src/main/java/com/rulecity/parse/data/ExternalNamesDefinition.java
@@ -1,5 +1,6 @@
 package com.rulecity.parse.data;
 
-public record ExternalNamesDefinition(String externalNameString, byte typeIndex)
+// typeIndex: 1-based OMF index per TIS 1.1 (variable-width: 1..32767, wider than a byte).
+public record ExternalNamesDefinition(String externalNameString, int typeIndex)
 {
 }
