+++ src/main/java/com/rulecity/parse/data/PublicNamesDefinition.java
@@ -2,7 +2,8 @@ package com.rulecity.parse.data;
 
 public record PublicNamesDefinition(String publicNameString,
         int publicOffset,
-        byte typeIndex
+        // 1-based OMF index per TIS 1.1 (variable-width: 1..32767, wider than a byte).
+        int typeIndex
 )
 {
 }
